services/similarity/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
sentence-transformers==3.4.1
torch==2.6.0
numpy==2.2.5
Pillow==11.2.1
ffmpeg-python==0.2.0
requests==2.32.3
